improve tests readability
